[fix] weapon repair

+++ include/weapon.h
@@ -156,8 +156,7 @@ i32 Weapon_remUses(const Weapon * wpn,
 //           OI BOSS COULD YOU SHARPEN MY SHIELD?
 
 
-void Weapon_Repair(Weapon * wpn, struct InvItem *item, u8 AP);
-
+void Weapon_Repair(Weapon * wpn, struct InvItem *item, i32 AP);
 
 
 #endif /* WEAPON_H */
